finance/finance-document-analyzer/pyproject.toml
[project]
name = "floagentdemo"
version = "0.1.0"
description = "floagentdemo using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pypdf==5.4.0",
    "python-multipart>=0.0.6",
    "streamlit==1.44.0",
    "fpdf2==2.8.2",
    "plotly==6.0.1",
    "kaleido==0.2.1"
]

[project.scripts]
floagentdemo = "floagentdemo.main:run"
run_crew = "floagentdemo.main:run"
train = "floagentdemo.main:train"
replay = "floagentdemo.main:replay"
test = "floagentdemo.main:test"
api = "floagentdemo.api:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
